enable multiple exports in package json
