+++ .gitignore
@@ -7,3 +7,5 @@ dist/
 main*
 module*
 demo
+parcel-bundle-reports
+pkg
