+++ package.json
@@ -2,27 +2,62 @@
   "name": "@sayari/trellis",
   "version": "0.7.0-rc.4",
   "description": "High-performance network visualization library, rendering to WebGL as well as other targets",
-  "source": "src/index.ts",
-  "main": "./main.js",
-  "module": "./module.js",
-  "types": "./index.d.ts",
+  "author": "James Conkling <james.lane.conkling@gmail.com> jameslaneconkling.github.io",
+  "license": "ISC",
   "scripts": {
-    "clean": "rm -rf dist && rm -rf .parcel-cache && mkdir dist",
+    "clean": "rm -rf dist && rm -rf .parcel-cache",
     "format": "prettier . --write",
     "test": "echo \"Error: no test specified\" && exit 1",
     "typecheck": "tsc --noEmit",
     "lint": "eslint \"src/**/*.ts\"",
     "lint:fix": "eslint \"src/**/**.{ts,tsx}\" --fix",
     "perf": "parcel perf/index.html",
-    "copy": "mv main* module* index.d* dist/",
+    "copy": "cp package* README.md dist/",
     "build:tsc": "tsc --project ./tsconfig.json",
-    "build": "npm run clean && parcel build && npm run copy && npm run build:tsc",
+    "build": "npm run clean && npm run build:tsc && parcel build",
+    "build:analyze": "npm run clean && parcel build --detailed-report --reporter @parcel/reporter-bundle-analyzer",
     "examples:dev": "npm run clean && parcel examples/index.html",
     "docs:dev": "npm run clean && parcel docs-src/index.html",
     "docs:build": "rm -rf ./docs/ && parcel build ./docs-src/index.html --dist-dir docs/ --public-url https://sayari-analytics.github.io/trellis/ && cp ./docs-src/assets/og-image.png ./docs/"
   },
-  "author": "James Conkling <james.lane.conkling@gmail.com> jameslaneconkling.github.io",
-  "license": "ISC",
+  "targets": {
+    "main": {
+      "source": "src/index.ts",
+      "outputFormat": "commonjs",
+      "isLibrary": true
+    },
+    "module": {
+      "source": "src/index.ts",
+      "outputFormat": "esmodule",
+      "isLibrary": true
+    }
+  },
+  "exports": {
+    ".": {
+      "import": "./module/index.js",
+      "require": "./main/index.js",
+      "types": "./src/index.d.ts"
+    },
+    "./react": {
+      "default": "./src/bindings/react/index.js",
+      "types": "./src/bindings/react/index.d.ts"
+    },
+    "./native": {
+      "default": "./src/bindings/native/index.js",
+      "types": "./src/bindings/native/index.d.ts"
+    },
+    "./webgl": {
+      "default": "./src/renderers/webgl/index.js",
+      "types": "./src/renderers/webgl/index.d.ts"
+    },
+    "./image": {
+      "default": "./src/renderers/image/index.js",
+      "types": "./src/renderers/image/index.d.ts"
+    }
+  },
+  "@parcel/resolver-default": {
+    "packageExports": true
+  },
   "dependencies": {
     "@pixi/unsafe-eval": "^6.5.5",
     "d3-color": "^3.1.0",
@@ -34,6 +69,7 @@
   },
   "devDependencies": {
     "@parcel/packager-ts": "^2.9.3",
+    "@parcel/reporter-bundle-analyzer": "^2.9.3",
     "@parcel/transformer-typescript-tsc": "^2.9.3",
     "@parcel/transformer-typescript-types": "^2.9.3",
     "@turf/clusters-kmeans": "^6.3.0",
@@ -68,14 +104,7 @@
     "ts-loader": "^6.2.2",
     "typescript": "^5.2.2"
   },
-  "keywords": [
-    "graph",
-    "network",
-    "infovis",
-    "visualization",
-    "react",
-    "webgl"
-  ],
+  "keywords": ["graph", "network", "infovis", "visualization", "react", "webgl"],
   "peerDependencies": {
     "react": ">=16.0"
   },
