+++ src/bindings/native/index.ts
@@ -1,0 +1,10 @@
+import Download from './download'
+import Selection from './selection'
+import Zoom from './zoom'
+
+import type { Options as DownloadOptions } from './download'
+import type { Options as ZoomOptions, ViewportChangeOptions } from './zoom'
+import type { Options as SelectionOptions, SelectionChangeEvent } from './selection'
+
+export { Download, Selection, Zoom }
+export type { DownloadOptions, SelectionOptions, ZoomOptions, ViewportChangeOptions, SelectionChangeEvent }
