+++ src/index.ts
@@ -1,95 +1,2 @@
-export * from './trellis'
-
-// layout
-import Cluster from './layout/cluster'
-import Collide from './layout/collide'
-import Components from './layout/components'
-import Fisheye from './layout/fisheye'
-import Force from './layout/force'
-import Hierarchy from './layout/hierarchy'
-import Radial from './layout/radial'
-
-import type { Options as CollideOptions } from './layout/collide'
-import type { Options as ComponentsOptions } from './layout/components'
-import type { Options as ForceOptions } from './layout/force'
-import type { Options as HierarchyOptions } from './layout/hierarchy'
-import type { Options as RadialOptions } from './layout/radial'
-
-export { Cluster, Collide, Components, Fisheye, Force, Hierarchy, Radial }
-export type { CollideOptions, ComponentsOptions, ForceOptions, HierarchyOptions, RadialOptions }
-
-// renderers
-import ImageRenderer, { BlobRenderer } from './renderers/image'
-import Renderer from './renderers/webgl'
-
-import type { Options as ImageOptions } from './renderers/image'
-import type {
-  NodePointerEvent,
-  NodeDragEvent,
-  EdgePointerEvent,
-  AnnotationPointerEvent,
-  AnnotationResizeEvent,
-  AnnotationDragEvent,
-  AnnotationResizePointerEvent,
-  ViewportPointerEvent,
-  ViewportDragEvent,
-  ViewportDragDecelerateEvent,
-  ViewportWheelEvent,
-  Options as RendererOptions
-} from './renderers/webgl'
-
-export { ImageRenderer, BlobRenderer, Renderer }
-export type {
-  ImageOptions,
-  NodePointerEvent,
-  NodeDragEvent,
-  EdgePointerEvent,
-  AnnotationPointerEvent,
-  AnnotationResizeEvent,
-  AnnotationDragEvent,
-  AnnotationResizePointerEvent,
-  ViewportPointerEvent,
-  ViewportDragEvent,
-  ViewportDragDecelerateEvent,
-  ViewportWheelEvent,
-  RendererOptions
-}
-
-// bindings
-import Download from './bindings/native/download'
-import Selection from './bindings/native/selection'
-import Zoom from './bindings/native/zoom'
-import ReactButton from './bindings/react/button'
-import ReactButtonGroup from './bindings/react/buttonGroup'
-import ReactRenderer from './bindings/react/renderer'
-import ReactSelection from './bindings/react/selection'
-import ReactZoom from './bindings/react/zoom'
-
-import type { Options as DownloadOptions } from './bindings/native/download'
-import type { Options as ZoomOptions, ViewportChangeOptions } from './bindings/native/zoom'
-import type { Options as SelectionOptions, SelectionChangeEvent } from './bindings/native/selection'
-import type { Props as ButtonProps } from './bindings/react/button'
-import type { Props as ButtonGroupProps } from './bindings/react/buttonGroup'
-import type { Props as RendererProps } from './bindings/react/renderer'
-import type { Props as ZoomProps } from './bindings/react/zoom'
-import type {
-  Props as SelectionProps,
-  ChildProps as SelectionChildProps,
-  SelectionChangeEvent as ReactSelectionChangeEvent
-} from './bindings/react/selection'
-
-export { Download, Selection, Zoom, ReactButton, ReactButtonGroup, ReactRenderer, ReactSelection, ReactZoom }
-export type {
-  DownloadOptions,
-  SelectionOptions,
-  ZoomOptions,
-  ViewportChangeOptions,
-  SelectionChangeEvent,
-  ButtonProps,
-  ButtonGroupProps,
-  RendererProps,
-  SelectionProps,
-  SelectionChildProps,
-  ZoomProps,
-  ReactSelectionChangeEvent
-}
+export * from './api'
+export * from './layout'
