+++ src/layout/force/index.ts
@@ -12,7 +12,7 @@ import {
   SimulationNodeDatum
 } from 'd3-force'
 import { Extend } from '../../types'
-import { Node, Edge } from '../../trellis'
+import { Node, Edge } from '../../api'
 
 export type Options = Partial<{
   nodeStrength: number
